feat: add new module ops
* refactor(lib.rs): move GitOps from client module to ops module
* feat(lib.rs): export MakeRelease and UpdateFromPr from ops module
* refactor(main.rs): update imports due to changes in lib.rs
* refactor(main.rs): replace client.changelog() with client.changelog_as_str() in debug and trace logs
* feat(git_ops.rs): add branch_status, branch_list, repo_status methods
* refactor(git_ops.rs): rename changelog method to changelog_as_str
* refactor(git_ops.rs): move git_ops.rs from client to ops directory
* feat(make_release.rs): add new file with MakeRelease trait and implementation
* feat(update_from_pr.rs): add new file with UpdateFromPr trait and implementation
* feat(mod.rs): add new file to export GitOps, MakeRelease, UpdateFromPr traits
* refactor(mod.rs): remove unused imports, functions and change visibility of some variables
* feat(mod.rs): add Error import from crate
* fix(mod.rs): rename changelog function to changelog_as_str
