+++ src/client/mod.rs
@@ -1,43 +1,27 @@
-use std::{
-    collections::HashMap,
-    env,
-    ffi::OsString,
-    fs,
-    path::{self},
-    str::FromStr,
-    sync::Arc,
-};
+use std::{collections::HashMap, env, ffi::OsString};
 
-mod git_ops;
 mod pull_request;
-
-pub use git_ops::GitOps;
 
 use self::pull_request::PullRequest;
 
-use chrono::{Datelike, NaiveDate, Utc};
 use config::Config;
 use git2::Repository;
-use keep_a_changelog::{
-    changelog::ChangelogBuilder, ChangeKind, Changelog, ChangelogParseOptions, Release, Version,
-};
-use octocrab::Octocrab;
-use url::Url;
+use keep_a_changelog::ChangeKind;
 
+use crate::Error;
 use crate::PrTitle;
-use crate::{release_notes_provider::ReleaseNotesProvider, Error};
 
 pub struct Client {
     #[allow(dead_code)]
-    settings: Config,
-    git_repo: Repository,
-    owner: String,
-    repo: String,
-    branch: Option<String>,
+    pub(crate) settings: Config,
+    pub(crate) git_repo: Repository,
+    pub(crate) owner: String,
+    pub(crate) repo: String,
+    pub(crate) branch: Option<String>,
     pull_request: Option<PullRequest>,
-    changelog: OsString,
-    changelog_update: Option<PrTitle>,
-    unreleased: Option<String>,
+    pub(crate) changelog: OsString,
+    pub(crate) changelog_update: Option<PrTitle>,
+    pub(crate) unreleased: Option<String>,
 }
 
 impl Client {
@@ -183,201 +167,6 @@ impl Client {
         self.branch == Some(default_branch)
     }
 
-    pub fn create_entry(&mut self) -> Result<(), Error> {
-        let mut pr_title = PrTitle::parse(self.title())?;
-        pr_title.pr_id = Some(self.pr_number());
-        pr_title.pr_url = Some(Url::from_str(self.pull_request())?);
-        pr_title.calculate_section_and_entry();
-
-        self.changelog_update = Some(pr_title);
-
-        Ok(())
-    }
-
-    pub fn update_changelog(&mut self) -> Result<Option<(ChangeKind, String)>, Error> {
-        log::debug!(
-            "Updating changelog: {:?} with entry {:?}",
-            self.changelog,
-            self.changelog_update
-        );
-
-        if self.changelog.is_empty() {
-            return Err(Error::NoChangeLogFileFound);
-        }
-
-        if let Some(update) = &mut self.changelog_update {
-            #[allow(clippy::needless_question_mark)]
-            return Ok(update.update_changelog(&self.changelog)?);
-        }
-        Ok(None)
-    }
-    /// Update the unreleased section to the changelog to `version`
-    pub fn update_unreleased(&mut self, version: &str) -> Result<(), Error> {
-        log::debug!(
-            "Updating unreleased section: {:?} with version {:?}",
-            self.changelog,
-            version,
-        );
-
-        if self.changelog.is_empty() {
-            return Err(Error::NoChangeLogFileFound);
-        }
-
-        if !version.is_empty() {
-            #[allow(clippy::needless_question_mark)]
-            return Ok(self.release_unreleased(version)?);
-        }
-
-        Ok(())
-    }
-
-    pub fn release_unreleased(&mut self, version: &str) -> Result<(), Error> {
-        let Some(log_file) = self.changelog.to_str() else {
-            return Err(Error::InvalidPath(self.changelog.to_owned()));
-        };
-
-        let repo_url =
-        //  match &self.pr_url {
-        //     Some(pr_url) => {
-        //         let url_string = pr_url.to_string();
-        //         let components = url_string.split('/').collect::<Vec<&str>>();
-        //         let url = format!("https://github.com/{}/{}", components[3], components[4]);
-        //         Some(url)
-        //     }
-        //     None => 
-        None
-        // }
-        ;
-
-        let mut change_log = if path::Path::new(log_file).exists() {
-            let file_contents = fs::read_to_string(path::Path::new(log_file))?;
-            log::trace!("file contents:\n---\n{}\n---\n\n", file_contents);
-            let options = if repo_url.is_some() {
-                Some(ChangelogParseOptions {
-                    url: repo_url.clone(),
-                    ..Default::default()
-                })
-            } else {
-                None
-            };
-
-            Changelog::parse_from_file(log_file, options)
-                .map_err(|e| Error::KeepAChangelog(e.to_string()))?
-        } else {
-            log::trace!("The changelog does not exist! Create a default changelog.");
-            let mut changelog = ChangelogBuilder::default()
-                .url(repo_url)
-                .build()
-                .map_err(|e| Error::KeepAChangelog(e.to_string()))?;
-            log::debug!("Changelog: {:#?}", changelog);
-            let release = Release::builder()
-                .build()
-                .map_err(|e| Error::KeepAChangelog(e.to_string()))?;
-            changelog.add_release(release);
-            log::debug!("Changelog: {:#?}", changelog);
-
-            changelog
-                .save_to_file(log_file)
-                .map_err(|e| Error::KeepAChangelog(e.to_string()))?;
-            changelog
-        };
-
-        // Get the unreleased section from the Changelog.
-        // If there is no unreleased section create it and add it to the changelog
-        let unreleased = if let Some(unreleased) = change_log.get_unreleased_mut() {
-            unreleased
-        } else {
-            let release = Release::builder()
-                .build()
-                .map_err(|e| Error::KeepAChangelog(e.to_string()))?;
-            change_log.add_release(release);
-            let unreleased = change_log.get_unreleased_mut().unwrap();
-            unreleased
-        };
-
-        let version = Version::parse(version).map_err(|e| Error::InvalidVersion(e.to_string()))?;
-        unreleased.set_version(version);
-
-        let today =
-            NaiveDate::from_ymd_opt(Utc::now().year(), Utc::now().month(), Utc::now().day());
-        if let Some(today) = today {
-            unreleased.set_date(today);
-        };
-
-        let unreleased_string = unreleased.to_string();
-        log::trace!("Release notes:\n\n---\n{}\n---\n\n", unreleased_string);
-        let _ = fs::write("release_notes.md", unreleased_string.clone());
-
-        self.unreleased = Some(unreleased_string);
-
-        change_log
-            .save_to_file(log_file)
-            .map_err(|e| Error::KeepAChangelog(e.to_string()))?;
-
-        Ok(())
-    }
-
-    pub async fn make_release(&self, version: &str) -> Result<(), Error> {
-        log::debug!("Making release {version}");
-
-        log::debug!("Creating octocrab instance {:?}", self.settings);
-        log::trace!(
-            "Creating octocrab for owner: {} and repo: {}",
-            self.owner(),
-            self.repo()
-        );
-
-        let opts = ChangelogParseOptions::default();
-        let changelog = match Changelog::parse_from_file(self.changelog(), Some(opts)) {
-            Ok(changelog) => changelog,
-            Err(e) => {
-                log::error!("Error parsing changelog: {e}");
-                return Err(Error::InvalidPath(self.changelog.clone()));
-            }
-        };
-
-        let release_notes = changelog.release_notes(version)?;
-        log::trace!("Release notes: {:#?}", release_notes);
-
-        let octocrab = match self.settings.get::<String>("pat") {
-            Ok(pat) => {
-                log::debug!("Using personal access token for authentication");
-                Arc::new(
-                    Octocrab::builder()
-                        .base_uri("https://api.github.com")?
-                        .personal_token(pat)
-                        .build()?,
-                )
-            }
-            // base_uri: https://api.github.com
-            // auth: None
-            // client: http client with the octocrab user agent.
-            Err(_) => {
-                log::debug!("Creating un-authenticated instance");
-                octocrab::instance()
-            }
-        };
-
-        let tag = format!("v{version}");
-        let commit = Client::get_commitish_for_tag(self, &octocrab, &tag).await?;
-        log::trace!("Commit: {:#?}", commit);
-
-        let release = octocrab
-            .repos(self.owner(), self.repo())
-            .releases()
-            .create(format!("v{version}").as_str())
-            .name(&release_notes.name)
-            .body(&release_notes.body)
-            .send()
-            .await?;
-
-        log::trace!("Release: {:#?}", release);
-
-        Ok(())
-    }
-}
-
-impl Client {
     pub fn section(&self) -> Option<&str> {
         if let Some(update) = &self.changelog_update {
             if let Some(section) = &update.section {
@@ -405,7 +194,7 @@ impl Client {
         }
     }
 
-    pub fn changelog(&self) -> &str {
+    pub fn changelog_as_str(&self) -> &str {
         if let Some(cl) = &self.changelog.to_str() {
             cl
         } else {
@@ -416,224 +205,4 @@ impl Client {
     pub fn set_changelog(&mut self, changelog: &str) {
         self.changelog = changelog.into();
     }
-
-    pub fn repo_status(&self) -> Result<String, Error> {
-        let statuses = self.git_repo.statuses(None)?;
-
-        log::trace!("Repo status length: {:?}", statuses.len());
-
-        let report = print_long(&statuses);
-        Ok(report)
-    }
-
-    pub fn branch_list(&self) -> Result<String, Error> {
-        let branches = self.git_repo.branches(None)?;
-
-        let mut output = String::from("\nList of branches:\n");
-        for item in branches {
-            let (branch, branch_type) = item?;
-            output = format!(
-                "{}\n# Branch and type: {:?}\t{:?}",
-                output,
-                branch.name(),
-                branch_type
-            );
-        }
-        output = format!("{}\n", output);
-
-        Ok(output)
-    }
-
-    pub fn branch_status(&self) -> Result<String, Error> {
-        let branch_remote = self.git_repo.find_branch(
-            format!("origin/{}", self.branch()).as_str(),
-            git2::BranchType::Remote,
-        )?;
-
-        if branch_remote.get().target() == self.git_repo.head()?.target() {
-            return Ok(format!(
-                "\n\nOn branch {}\nYour branch is up to date with `{}`\n",
-                self.branch(),
-                branch_remote.name()?.unwrap()
-            ));
-        }
-
-        let local = self.git_repo.head()?.target().unwrap();
-        let remote = branch_remote.get().target().unwrap();
-
-        let (ahead, behind) = self.git_repo.graph_ahead_behind(local, remote)?;
-
-        let output = format!(
-            "Your branch is {} commits ahead and {} commits behind\n",
-            ahead, behind
-        );
-
-        Ok(output)
-    }
-}
-
-// This function print out an output similar to git's status command in long
-// form, including the command-line hints.
-fn print_long(statuses: &git2::Statuses) -> String {
-    let mut header = false;
-    let mut rm_in_workdir = false;
-    let mut changes_in_index = false;
-    let mut changed_in_workdir = false;
-
-    let mut output = String::new();
-
-    // Print index changes
-    for entry in statuses
-        .iter()
-        .filter(|e| e.status() != git2::Status::CURRENT)
-    {
-        if entry.status().contains(git2::Status::WT_DELETED) {
-            rm_in_workdir = true;
-        }
-        let istatus = match entry.status() {
-            s if s.contains(git2::Status::INDEX_NEW) => "new file: ",
-            s if s.contains(git2::Status::INDEX_MODIFIED) => "modified: ",
-            s if s.contains(git2::Status::INDEX_DELETED) => "deleted: ",
-            s if s.contains(git2::Status::INDEX_RENAMED) => "renamed: ",
-            s if s.contains(git2::Status::INDEX_TYPECHANGE) => "typechange:",
-            _ => continue,
-        };
-        if !header {
-            output = format!(
-                "{}\n\
-                # Changes to be committed:
-                #   (use \"git reset HEAD <file>...\" to unstage)
-                #",
-                output
-            );
-            header = true;
-        }
-
-        let old_path = entry.head_to_index().unwrap().old_file().path();
-        let new_path = entry.head_to_index().unwrap().new_file().path();
-        match (old_path, new_path) {
-            (Some(old), Some(new)) if old != new => {
-                output = format!(
-                    "{}\n#\t{}  {} -> {}",
-                    output,
-                    istatus,
-                    old.display(),
-                    new.display()
-                );
-            }
-            (old, new) => {
-                output = format!(
-                    "{}\n#\t{}  {}",
-                    output,
-                    istatus,
-                    old.or(new).unwrap().display()
-                );
-            }
-        }
-    }
-
-    if header {
-        changes_in_index = true;
-        output = format!("{}\n", output);
-    }
-    header = false;
-
-    // Print workdir changes to tracked files
-    for entry in statuses.iter() {
-        // With `Status::OPT_INCLUDE_UNMODIFIED` (not used in this example)
-        // `index_to_workdir` may not be `None` even if there are no differences,
-        // in which case it will be a `Delta::Unmodified`.
-        if entry.status() == git2::Status::CURRENT || entry.index_to_workdir().is_none() {
-            continue;
-        }
-
-        let istatus = match entry.status() {
-            s if s.contains(git2::Status::WT_MODIFIED) => "modified: ",
-            s if s.contains(git2::Status::WT_DELETED) => "deleted: ",
-            s if s.contains(git2::Status::WT_RENAMED) => "renamed: ",
-            s if s.contains(git2::Status::WT_TYPECHANGE) => "typechange:",
-            _ => continue,
-        };
-
-        if !header {
-            output = format!(
-                "{}\n# Changes not staged for commit:\n#   (use \"git add{} <file>...\" to update what will be committed)\n#   (use \"git checkout -- <file>...\" to discard changes in working directory)\n#               ",
-                output,
-                if rm_in_workdir { "/rm" } else { "" }
-            );
-            header = true;
-        }
-
-        let old_path = entry.index_to_workdir().unwrap().old_file().path();
-        let new_path = entry.index_to_workdir().unwrap().new_file().path();
-        match (old_path, new_path) {
-            (Some(old), Some(new)) if old != new => {
-                output = format!(
-                    "{}\n#\t{}  {} -> {}",
-                    output,
-                    istatus,
-                    old.display(),
-                    new.display()
-                );
-            }
-            (old, new) => {
-                output = format!(
-                    "{}\n#\t{}  {}",
-                    output,
-                    istatus,
-                    old.or(new).unwrap().display()
-                );
-            }
-        }
-    }
-
-    if header {
-        changed_in_workdir = true;
-        output = format!("{}\n#\n", output);
-    }
-    header = false;
-
-    // Print untracked files
-    for entry in statuses
-        .iter()
-        .filter(|e| e.status() == git2::Status::WT_NEW)
-    {
-        if !header {
-            output = format!(
-                "{}# Untracked files\n#   (use \"git add <file>...\" to include in what will be committed)\n#",
-                output
-            );
-            header = true;
-        }
-        let file = entry.index_to_workdir().unwrap().old_file().path().unwrap();
-        output = format!("{}\n#\t{}", output, file.display());
-    }
-    header = false;
-
-    // Print ignored files
-    for entry in statuses
-        .iter()
-        .filter(|e| e.status() == git2::Status::IGNORED)
-    {
-        if !header {
-            output = format!(
-                "{}\n# Ignored files\n#   (use \"git add -f <file>...\" to include in what will be committed)\n#",
-                output
-            );
-            header = true;
-        }
-        let file = entry.index_to_workdir().unwrap().old_file().path().unwrap();
-        output = format!("{}\n#\t{}", output, file.display());
-    }
-
-    if !changes_in_index && changed_in_workdir {
-        output = format!(
-            "{}\n
-            no changes added to commit (use \"git add\" and/or \
-            \"git commit -a\")",
-            output
-        );
-    }
-
-    output
 }
