+++ src/lib.rs
@@ -1,9 +1,12 @@
 mod client;
 mod error;
+mod ops;
 mod pr_title;
 mod release_notes_provider;
 
 pub use client::Client;
-pub use client::GitOps;
 pub use error::Error;
+pub use ops::GitOps;
+pub use ops::MakeRelease;
+pub use ops::UpdateFromPr;
 pub use pr_title::PrTitle;
