+++ src/ops/mod.rs
@@ -1,0 +1,7 @@
+mod git_ops;
+mod make_release;
+mod update_from_pr;
+
+pub use git_ops::GitOps;
+pub use make_release::MakeRelease;
+pub use update_from_pr::UpdateFromPr;
